FInished Event Space method
paralleloperation still has wrong scale, i did some research into multiple failures, and its not something that should be considered, given that if there are multiple failures, they will be recorded as the nth failure of the component. analytical parallel is done. ESM has been checked for a problem with 15 nodes, and has frozen. there are a bunch of optimizations we could do here, but i'll look into the literature, given that ESM may just be really poor for "larger" systems